Address comments (#21225)
