--- test/tests/bulk_material/tests
@@ -1,9 +1,0 @@
-[Tests]
-  [test]
-    type = RunApp
-    input = test.i
-    design = BulkMaterial.md
-    issues = '#21225'
-    requirement = 'The bulk material system shall provide the capability for gathering and serializing variable and material property values for bulk property computation.'
-  []
-[]
